ltp: sharding: remove the tmp file before creation
When running multiple tests in different test sections in the LAVA job
definition file, the alltests file was only adding more and more tests
to the file. Which lead to that the first tests that was added was
running in all the following tests as well.

Ensuring to delete the 'alltests' file before generating the file, will
fix the problem of running the same set of tests in the following test
runs.

Reported-by: Naresh Kamboju <naresh.kamboju@linaro.org>
Signed-off-by: Anders Roxell <anders.roxell@linaro.org>