Update rt tests pi stress and cyclicdeadline (#567)
* automated: linux: pi-stress: fix threshold

The pi-stress results shouldn't be based on the minimum value, it should
be -X% from the max_inversion.
Anything less than threshold is a fail.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>

* automated: linux: pi-stress: allow user-specified baseline

Previously, the pi-stress script dynamically computed the baseline
(minimum inversion value) across all iterations. This patch adds support
for a new `-x` option that lets users manually specify a fixed
`user_baseline` value.

Using a fixed baseline is preferred over percentage-based thresholds in
this context, as it offers more consistent and predictable
evaluation—especially in controlled benchmarking environments or when
comparing results across different systems and runs. Percentages can be
misleading if the dynamic baseline fluctuates due to noise or system
conditions.

This change improves flexibility and repeatability in post-processing
and performance analysis.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>

* automated: linux: cyclicdeadline: add user-specified baseline

Previously, the cyclicdeadline script dynamically computed the minimum
latency value across all iterations. This patch introduces a new `-x`
option that enables users to specify a fixed `user_baseline` value
instead.

Using a fixed baseline is preferred over percentage-based thresholds in
this context, as it offers more consistent and predictable
evaluation—especially in controlled benchmarking environments or when
comparing results across different systems and runs. Percentages can be
misleading if the dynamic baseline fluctuates due to noise or system
conditions.

This change improves flexibility and repeatability in post-processing
and performance analysis.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>

---------

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>